refactor(admin-vue3): 重构字典管理页面，优化交互方式
- 新增 Data 和 DictGroup 组件
- 优化字典数据列表和字典类型对话框的结构，初步使用树形图展示字典项
- 移除冗余代码，提高代码可读性和维护性
