+++ admin-vue3/src/views/system/dict/components/data.vue
